refactor data fetching with hooks, folder re-org
